document ComputedStates
